Fixing bug causing slow OAI-PMH output.
Records were being added to sets multiple times and never being removed.
This was masked by use of :uniq, which resolved the issue using
"DISTINCT" in the SQL.

The record/set membership logic is fixed in this commit. The number of
queries is also being reduced by using default_scope to eagerly fetch
:sets and :underlying_concept in all Record queries.

Also adding profiling support in development.